move migratedFixtures to dist
