+++ packages/ricos-content/web/src/converters/draft/fromDraft/fromDraft.spec.ts
@@ -5,8 +5,8 @@ import { compare } from '../../../comparision/compare';
 import fixture from '../../../../../../../e2e/tests/fixtures/intro.json';
 import complexFixture from '../../../../../../../e2e/tests/fixtures/migration-content.json';
 import { getTextNodes } from './getTextNodes';
-import ricosFixture from './migratedFixtures/intro.json';
-import complexRicosFixture from './migratedFixtures/migration-content.json';
+import ricosFixture from '../../../../statics/json/migratedFixtures/intro.json';
+import complexRicosFixture from '../../../../statics/json/migratedFixtures/migration-content.json';
 import { Node_Type, Decoration_Type, RichContent } from 'ricos-schema';
 import { convertBlockDataToRicos } from './convertRicosPluginData';
 import { IMAGE_TYPE } from '../../../consts';
